% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{chrome_launcher}
\alias{chrome_launcher}
\title{Launch Chrome at a given window size}
\usage{
chrome_launcher(width = 1440, height = 900, app_mode = TRUE)
}
\arguments{
\item{width, height}{window size in pixels}

\item{app_mode}{use Chrome's app mode (no tabs)}
}
\description{
Launch Chrome at a given window size
}
