% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{convert_symbol_to_ensembl}
\alias{convert_symbol_to_ensembl}
\title{Convert Gene Symbols to Ensembl IDs}
\usage{
convert_symbol_to_ensembl(gene_symbols, species = "Homo sapiens")
}
\arguments{
\item{gene_symbols}{Character vector of gene symbols}

\item{species}{Character string}
}
\value{
Named character vector
}
\description{
Convert Gene Symbols to Ensembl IDs
}
