% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{get_orgdb}
\alias{get_orgdb}
\title{Get OrgDb object from species name}
\usage{
get_orgdb(species)
}
\arguments{
\item{species}{Character string: either "Homo sapiens" or "Mus musculus"}
}
\value{
Corresponding OrgDb object
}
\description{
Returns the appropriate OrgDb object for the given species.
}
