% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{get_reactome_code}
\alias{get_reactome_code}
\title{Get Reactome Code from species name}
\usage{
get_reactome_code(species)
}
\arguments{
\item{species}{Character string: either "Homo sapiens" or "Mus musculus"}
}
\value{
Character Reactome organism code
}
\description{
Returns the Reactome organism code for a given species.
}
