% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{is_ensembl_id}
\alias{is_ensembl_id}
\title{Check if IDs are Ensembl}
\usage{
is_ensembl_id(ids)
}
\arguments{
\item{ids}{Character vector of gene identifiers}
}
\value{
Logical indicating whether majority of IDs resemble Ensembl format
}
\description{
Checks whether a vector of gene IDs are mostly Ensembl IDs.
}
