% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{is_symbol}
\alias{is_symbol}
\title{Check if IDs are gene symbols}
\usage{
is_symbol(ids)
}
\arguments{
\item{ids}{Character vector of gene identifiers}
}
\value{
Logical indicating whether IDs are mostly gene symbols
}
\description{
Heuristically determines whether gene IDs appear to be symbols rather than Ensembl.
}
