% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cancer_gene_census.R
\name{mod_cancer_gene_census}
\alias{mod_cancer_gene_census}
\title{Module to visualize overlapping genes from DE analysis with Cancer Gene Census}
\usage{
mod_cancer_gene_census(input, output, session, res_reactive, filtered_data_rv)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{res_reactive}{Reactive expression containing DE results (e.g., logFC, padj, etc.)}
}
\description{
This module compares differentially expressed genes (DEGs) from the DE results with a
Cancer Gene Census dataset, visualizes the overlap using a Venn diagram, and lists
the overlapping genes in a table.
}
