% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cross_plot.R
\name{mod_cross_plot}
\alias{mod_cross_plot}
\title{Cross Plot and Venn Diagram Module}
\usage{
mod_cross_plot(input, output, session, data, dds_rv)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{data}{Data containing counts, samples, norm_counts, and species}

\item{dds_rv}{A reactive value containing the DESeq2 object.}
}
\value{
None. Updates Cross Plot and Venn Diagram and download handlers.
}
\description{
Cross Plot and Venn Diagram Module
}
