% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cross_tab_ui.R
\name{mod_cross_tab_ui}
\alias{mod_cross_tab_ui}
\title{Cross Plot tabPanel (UI module)}
\usage{
mod_cross_tab_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Cross Plot tabPanel (UI module)
}
