% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{mod_data_upload_design}
\alias{mod_data_upload_design}
\title{Data upload and design}
\usage{
mod_data_upload_design(input, output, session, data, dds_rv)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{data}{ReactiveValues to store processed data (counts, samples, norm_counts).}

\item{dds_rv}{A reactiveVal container for storing the DESeq2 dataset.}
}
\value{
A list with elements: counts, samples, norm_counts, dds
}
\description{
Load and process uploaded count and design files into DESeq2 objects
}
