% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_deseq.R
\name{mod_differential_expression}
\alias{mod_differential_expression}
\title{Differential Expression Module}
\usage{
mod_differential_expression(
  input,
  output,
  session,
  filtered_data_rv,
  filtered_dds_rv,
  res_reactive
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{filtered_data_rv}{a reactive list with counts, samples, norm_counts, species}

\item{filtered_dds_rv}{a reactive valaue holding DESeqDataSet}

\item{res_reactive}{ReactiveVal to store DESeq2 result}
}
\value{
None (outputs rendered)
}
\description{
Handles DESeq2-based differential expression, renders results table, and provides a clustered heatmap of top genes
}
