% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_easy_comparison_builder.R
\name{mod_easy_compare_server}
\alias{mod_easy_compare_server}
\title{Easy Comparison Builder Server}
\usage{
mod_easy_compare_server(id, filtered_data_rv)
}
\arguments{
\item{id}{module id}

\item{filtered_data_rv}{reactiveValues; must contain $samples (data.frame with rownames as sample IDs)}
}
\value{
list of reactives (see code)
}
\description{
Uses filtered_data_rv$samples to define Test vs Reference comparisons
(simple, pooled, paired) and optional covariates to adjust.
}
