% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_easy_compare_ui.R
\name{mod_easy_compare_ui}
\alias{mod_easy_compare_ui}
\title{Easy Comparison Builder UI}
\usage{
mod_easy_compare_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Build Test vs Reference comparisons without contrasts.
}
