% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gene_expression.R
\name{mod_gene_expression_plot}
\alias{mod_gene_expression_plot}
\title{Module to visualize individual gene expression across sample groups}
\usage{
mod_gene_expression_plot(input, output, session, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{ReactiveValues containing counts, samples, norm_counts, and species}
}
\value{
None. Outputs are rendered to the UI
}
\description{
This module creates a boxplot (with jittered points) of log2 normalized gene expression
values for a user-input gene or genes, grouped by a selected metadata variable.
}
