% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utils.R
\name{mod_geo_loader}
\alias{mod_geo_loader}
\title{GEO Loader Module}
\usage{
mod_geo_loader(input, output, session, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{ReactiveValues storing counts, samples, norm_counts, species}
}
\description{
Attempts to download raw counts and metadata from a GEO accession.
Checks if valid GEO ID, counts, and metadata are available.
}
