% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goi_heatmap.R
\name{mod_goi_heatmap}
\alias{mod_goi_heatmap}
\title{Genes of Interest Heatmap Module}
\usage{
mod_goi_heatmap(input, output, session, filtered_data_rv)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{filtered_data_rv}{A reactiveValues list containing:
\itemize{
\item \code{norm_counts}: normalized gene expression matrix (genes x samples)
\item \code{samples}: metadata for samples
\item \code{species}: species name
}}
}
\description{
Allows the user to upload a gene list and visualize expression as a heatmap
using \code{filtered_data_rv$norm_counts}.
}
