% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gsea.R
\name{mod_gsea_server}
\alias{mod_gsea_server}
\title{GSEA Module (server)}
\usage{
mod_gsea_server(id, filtered_data_rv, res_reactive, de_sel = NULL)
}
\arguments{
\item{id}{module id (must match mod_gsea_tab_ui)}

\item{filtered_data_rv}{reactiveValues list containing at least $species}

\item{res_reactive}{reactiveVal containing DE results (data.frame)}

\item{de_sel}{(optional) list returned by mod_de_server(), providing ref/test levels for filenames}
}
\description{
Runs Gene Set Enrichment Analysis (GSEA) using clusterProfiler and msigdbr,
supports multiple databases (GO, KEGG, Reactome, Hallmark) and visualizes
the top enriched pathways with dot plots and enrichment scores.
}
