% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gsva.R
\name{mod_gsva_server}
\alias{mod_gsva_server}
\title{GSVA / ssGSEA module Server}
\usage{
mod_gsva_server(
  id,
  dds_rv,
  filtered_data_rv,
  group_var_react = NULL,
  ref_level_react = NULL,
  test_level_react = NULL
)
}
\arguments{
\item{id}{Shiny module ID}

\item{dds_rv}{reactiveVal returning a DESeqDataSet}

\item{filtered_data_rv}{ReactiveValues with at least $species}

\item{group_var_react}{optional reactive() for DE tab "Variable to test"}

\item{ref_level_react}{optional reactive() for DE tab "Reference Condition"}

\item{test_level_react}{optional reactive() for DE tab "Test Condition"}

\item{de_sel}{(optional) list returned by mod_de_server(), providing ref/test levels for filenames and defaults}
}
\description{
GSVA / ssGSEA module Server
}
