% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gsva.R
\name{mod_gsva_ui}
\alias{mod_gsva_ui}
\title{GSVA / ssGSEA module UI (uses DE tab selections by default)}
\usage{
mod_gsva_ui(id)
}
\arguments{
\item{id}{Shiny module ID}
}
\description{
GSVA / ssGSEA module UI (uses DE tab selections by default)
}
