% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_log.R
\name{mod_logger_server}
\alias{mod_logger_server}
\title{Logger Module Server}
\usage{
mod_logger_server(id, input_all)
}
\arguments{
\item{id}{Shiny module id}

\item{input_all}{The full Shiny input object (pass the top-level \code{input})}
}
\value{
reactiveVal character() with log entries
}
\description{
Logger Module Server
}
