% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_log.R
\name{mod_logger_ui}
\alias{mod_logger_ui}
\title{Logger Module UI}
\usage{
mod_logger_ui(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
A Shiny UI element to display the log and a download button
}
\description{
Logger Module UI
}
