% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nonverlap_path.R
\name{mod_nonoverlap_server}
\alias{mod_nonoverlap_server}
\title{Non-overlap Pathway Analysis Server}
\usage{
mod_nonoverlap_server(
  id,
  filtered_data_rv,
  pathway_result_rv,
  geneList_rv,
  geneListU_rv
)
}
\arguments{
\item{id}{module id}

\item{filtered_data_rv}{reactiveValues with $species}

\item{pathway_result_rv}{reactiveVal holding the \emph{initial} enrichment result (enrichResult)}

\item{geneList_rv}{reactiveVal named numeric vector (log2FC), names = ENTREZID}

\item{geneListU_rv}{reactiveVal to store the updated non-overlap gene list (named by ENTREZID)}
}
\description{
Removes genes already present in current enrichment results and re-runs
enrichment on the remaining (non-overlap) set.
}
