% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nonverlap_tab_ui.R
\name{mod_nonoverlap_tab_ui}
\alias{mod_nonoverlap_tab_ui}
\title{Non-overlap Pathway tabPanel (UI)}
\usage{
mod_nonoverlap_tab_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Non-overlap Pathway tabPanel (UI)
}
