% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pathway_analysis.R
\name{mod_pathway_analysis}
\alias{mod_pathway_analysis}
\title{Pathway Analysis Module}
\usage{
mod_pathway_analysis(
  input,
  output,
  session,
  data,
  res_reactive,
  geneList_rv,
  kegg_pathway_results,
  d1_merged_rv,
  pathway_input_rv
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{ReactiveValues containing counts, norm_counts, samples, species}

\item{res_reactive}{ReactiveVal holding DESeq2 results}

\item{geneList_rv}{ReactiveVal for log2FC vector}

\item{kegg_pathway_results}{ReactiveVal for KEGG pathway analysis result}

\item{d1_merged_rv}{Reactive Data frame containing gene symbols, logFC, and padj}

\item{pathway_input_rv}{Reactive pathfindR input data frame}
}
\description{
This module handles pathway enrichment analysis using clusterProfiler and visualizations using enrichplot and pathfindR.
}
