% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nonverlap_path.R
\name{mod_pathway_analysis_non_overlap}
\alias{mod_pathway_analysis_non_overlap}
\title{Non-overlapping Pathway Analysis Module}
\usage{
mod_pathway_analysis_non_overlap(
  input,
  output,
  session,
  geneList_rv,
  filtered_data_rv,
  pathway_result_rv,
  res_reactive,
  geneListU_rv
)
}
\arguments{
\item{input}{Shiny input object containing the user interface input values.}

\item{output}{Shiny output object to render the pathway analysis results and visualizations.}

\item{session}{Shiny session object for managing the session and UI updates.}

\item{geneList_rv}{Reactive object containing the initial list of genes, typically with log2 fold change values and corresponding Entrez IDs, used for the first pathway analysis.}

\item{filtered_data_rv}{Reactive object containing the filtered data, which includes normalized counts, samples, and species information.}

\item{pathway_result_rv}{Reactive object containing the initial pathway analysis results from clusterProfiler, including gene IDs and enrichment terms.}

\item{res_reactive}{Reactive object containing the DESeq2 results (log2 fold change, p-value, and adjusted p-value).}

\item{geneListU_rv}{Reactive object containing the unique list of genes, with log2 fold change values and corresponding Entrez IDs, used for the non-overlapping pathway analysis.}
}
\description{
This module handles the pathway enrichment analysis for non-overlapping genes. It filters out the genes that are already included in the initial pathway results and reruns the pathway analysis using clusterProfiler.
}
