% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pathway_plots_ui.R
\name{mod_pathway_plots_ui}
\alias{mod_pathway_plots_ui}
\title{Pathway Plots UI (tabbed: heatmap / tree / upset)}
\usage{
mod_pathway_plots_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Pathway Plots UI (tabbed: heatmap / tree / upset)
}
