% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pca_server.R
\name{mod_pca_server}
\alias{mod_pca_server}
\title{PCA Cluster Server}
\usage{
mod_pca_server(id, filtered_data_rv)
}
\arguments{
\item{id}{module id}

\item{filtered_data_rv}{reactiveValues with $norm_counts, $samples, $species}
}
\description{
PCA Cluster Server
}
