% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qc_plot.R
\name{mod_qc_plot}
\alias{mod_qc_plot}
\title{Quality Control Plot Module}
\usage{
mod_qc_plot(input, output, session, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{ReactiveValues containing counts, samples, norm_counts, and species}
}
\value{
None. Outputs are rendered to UI.
}
\description{
Generates various QC plots such as PCA, sample distance heatmap, mean-variance plots,
and gene expression variance histograms from normalized RNA-seq count data.
}
