% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_select_dynamic.R
\name{mod_sample_select_dynamic}
\alias{mod_sample_select_dynamic}
\title{Sample Selection Module (Dynamic)}
\usage{
mod_sample_select_dynamic(
  input,
  output,
  session,
  dds_rv,
  loaded_data_rv,
  filtered_data_rv,
  filtered_dds_rv
)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dds_rv}{reactiveVal containing the full DESeq2 object.}

\item{loaded_data_rv}{reactiveValues containing loaded raw data.}

\item{filtered_data_rv}{reactiveVal for filtered dataset.}

\item{filtered_dds_rv}{reactiveVal for filtered DESeq2 object.}
}
\description{
This module dynamically creates sample filtering UI elements based on the metadata
and allows filtering of samples for downstream analyses.
}
