% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tf_tab_ui.R
\name{mod_tf_tab_ui}
\alias{mod_tf_tab_ui}
\title{Transcription Factor Enrichment UI (content-only)}
\usage{
mod_tf_tab_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Transcription Factor Enrichment UI (content-only)
}
