% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{convert_ensembl_to_symbol}
\alias{convert_ensembl_to_symbol}
\title{Convert Ensembl IDs to Gene Symbols}
\usage{
convert_ensembl_to_symbol(gene_ids, species = "Homo sapiens")
}
\arguments{
\item{gene_ids}{Character vector of Ensembl IDs}

\item{species}{Character string}
}
\value{
Named character vector
}
\description{
Convert Ensembl IDs to Gene Symbols
}
