% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_de_tab_ui.R
\name{mod_de_tab_ui}
\alias{mod_de_tab_ui}
\title{Differential Expression tabPanel (UI module)}
\usage{
mod_de_tab_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Differential Expression tabPanel (UI module)
}
