% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_deseq.R
\name{mod_differential_expression}
\alias{mod_differential_expression}
\title{Differential Expression Module}
\usage{
mod_differential_expression(input, output, session, data, dds_rv, res_reactive)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{ReactiveValues with counts, samples, norm_counts, species}

\item{dds_rv}{ReactiveVal holding DESeqDataSet}

\item{res_reactive}{ReactiveVal to store DESeq2 result}
}
\value{
None (outputs rendered)
}
\description{
Handles DESeq2-based differential expression, renders results table, and provides a clustered heatmap of top genes
}
