% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tf.R
\name{mod_tf_enrichment_analysis}
\alias{mod_tf_enrichment_analysis}
\title{Transcription Factor Enrichment Analysis Module}
\usage{
mod_tf_enrichment_analysis(
  input,
  output,
  session,
  res_reactive,
  filtered_data_rv
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{res_reactive}{Reactive data with DE results}

\item{filtered_data_rv}{Reactive list containing raw data}
}
\description{
This module allows users to perform transcription factor enrichment analysis
based on selected datasets stored locally (gzipped) in inst/extdata.
}
