% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tf.R
\name{mod_tf_enrichment_server}
\alias{mod_tf_enrichment_server}
\title{TF Enrichment Server}
\usage{
mod_tf_enrichment_server(id, res_reactive, filtered_data_rv, cmp = NULL)
}
\arguments{
\item{id}{Shiny module id}

\item{res_reactive}{reactiveVal with DE results}

\item{filtered_data_rv}{reactiveValues with at least $species}

\item{cmp}{Optional comparison bridge from make_cmp_bridge(); used for titles/filenames}
}
\description{
TF Enrichment Server
}
